requirements.txt
python-telegram-bot[job-queue]==21.3
python-dotenv==1.0.1
aiosqlite==0.20.0
